+++ examples/actor.rs
@@ -1,7 +1,8 @@
 use std::error::Error;
 use std::process::exit;
+use std::thread::sleep;
 use std::time::Duration;
-use tyra::prelude::{Actor, ActorContext, ActorFactory, ActorMessage, ActorResult, ActorSystem, Handler, TyraConfig};
+use tyra::prelude::{Actor, ActorContext, ActorError, ActorFactory, ActorMessage, ActorResult, ActorSystem, Handler, TyraConfig};
 
 struct MessageA {
     text: String,
@@ -24,7 +25,18 @@ struct HelloWorld {
     count: usize,
 }
 
-impl Actor for HelloWorld {}
+impl Actor for HelloWorld {
+    fn on_error(&mut self, _context: &ActorContext<Self>, err: Box<dyn Error>) -> ActorResult {
+        println!("{:?}", err);
+        if err.is::<ActorError>() {
+            return ActorResult::Ok;
+        }
+        if err.is::<std::io::Error>() {
+            return ActorResult::Stop;
+        }
+        return ActorResult::Kill;
+    }
+}
 
 struct HelloWorldFactory {
     text: String,
@@ -40,20 +52,20 @@ impl ActorFactory<HelloWorld> for HelloWorldFactory {
     }
 }
 impl Handler<MessageA> for HelloWorld {
-    fn handle(&mut self, msg: MessageA, _context: &ActorContext<Self>) -> ActorResult {
+    fn handle(&mut self, msg: MessageA, _context: &ActorContext<Self>) -> Result<ActorResult, Box<dyn Error>> {
         let text: String = [self.text.clone(), String::from(msg.text)].join(" -> ");
         self.count += 1;
         println!("AAAA: {} Count: {}", text, self.count);
-        ActorResult::Ok
+        return Ok(ActorResult::Ok);
     }
 }
 
 impl Handler<MessageB> for HelloWorld {
-    fn handle(&mut self, msg: MessageB, _context: &ActorContext<Self>) -> ActorResult {
+    fn handle(&mut self, msg: MessageB, _context: &ActorContext<Self>) -> Result<ActorResult, Box<dyn Error>> {
         let text: String = [self.text.clone(), String::from(msg.text)].join(" -> ");
         self.count -= 1;
         println!("BBBB: {} Count: {}", text, self.count);
-        ActorResult::Ok
+        return Ok(ActorResult::Ok);
     }
 }
 
@@ -74,27 +86,28 @@ fn main() {
         .set_pool_name("aye")
         .spawn("hello-world", hw)
         .unwrap();
-    x.send(MessageA {
+    let _ = x.send(MessageA {
         text: String::from("sers+1"),
     });
-    x.send(MessageA {
+    let _ = x.send(MessageA {
         text: String::from("sers+2"),
     });
-    x.send(MessageB {
+    let _ = x.send(MessageB {
         text: String::from("sers-1"),
     });
-    x.send(MessageA {
+    let _ = x.send(MessageA {
         text: String::from("sers+3"),
     });
-    x.send(MessageA {
+    let _ = x.send(MessageA {
         text: String::from("sers+4"),
     });
-    x.send(MessageA {
+    let _ = x.send(MessageA {
         text: String::from("sers+5"),
     });
 
     //x.send(MessageUnsupported{text: String::from("sers")});
 
+    sleep(Duration::from_millis(125));
     actor_system.stop(Duration::from_secs(1));
     exit(actor_system.await_shutdown());
 }