+++ examples/getting_started.rs
@@ -54,9 +54,9 @@ impl ActorFactory<TemplateActor> for TemplateActorFactory {
 //Handler//
 ///////////
 impl Handler<TemplateMessage> for TemplateActor {
-    fn handle(&mut self, _msg: TemplateMessage, _context: &ActorContext<Self>) -> ActorResult {
+    fn handle(&mut self, _msg: TemplateMessage, _context: &ActorContext<Self>) -> Result<ActorResult, Box<dyn Error>> {
         println!("SERS: {}", _msg.id);
-        ActorResult::Ok
+        Ok(ActorResult::Ok)
     }
 }