+++ src/actor/handler.rs
@@ -1,3 +1,4 @@
+use std::error::Error;
 use crate::actor::actor::Actor;
 use crate::actor::context::ActorContext;
 use crate::message::actor_message::ActorMessage;
@@ -13,6 +14,7 @@ use crate::prelude::{ActorResult, BulkActorMessage, SerializedMessage};
 /// Basic usage:
 ///
 /// ```rust
+/// use std::error::Error;
 /// use tyra::prelude::{TyraConfig, ActorSystem, ActorFactory, ActorContext, SerializedMessage, ActorMessage, Handler, Actor, ActorResult};
 ///
 /// struct TestActor {}
@@ -22,8 +24,8 @@ use crate::prelude::{ActorResult, BulkActorMessage, SerializedMessage};
 /// impl ActorMessage for FooBar {}
 ///
 /// impl Handler<FooBar> for TestActor {
-///     fn handle(&mut self, _msg: FooBar, _context: &ActorContext<Self>) -> ActorResult {
-///         ActorResult::Ok
+///     fn handle(&mut self, _msg: FooBar, _context: &ActorContext<Self>) -> Result<ActorResult, Box<dyn Error>> {
+///         Ok(ActorResult::Ok)
 ///     }
 /// }
 /// ```
@@ -32,19 +34,23 @@ where
     Self: Actor + Sized,
     M: ActorMessage,
 {
-    fn handle(&mut self, msg: M, context: &ActorContext<Self>) -> ActorResult;
+    fn handle(&mut self, msg: M, context: &ActorContext<Self>) -> Result<ActorResult, Box<dyn Error>>;
 }
 
 impl<A> Handler<ActorStopMessage> for A
 where
     A: Actor + Sized,
 {
-    fn handle(&mut self, _msg: ActorStopMessage, context: &ActorContext<A>) -> ActorResult {
+    fn handle(&mut self, _msg: ActorStopMessage, context: &ActorContext<A>) -> Result<ActorResult, Box<dyn Error>> {
         let actor_result = self.on_actor_stop(context);
-        if actor_result != ActorResult::Stop || actor_result != ActorResult::Kill {
-            return ActorResult::Stop;
+        if actor_result.is_err() {
+            return actor_result;
         }
-        return actor_result;
+        let actor_result = actor_result.unwrap();
+        if actor_result != ActorResult::Stop || actor_result != ActorResult::Kill {
+            return Ok(ActorResult::Stop);
+        }
+        return Ok(actor_result);
     }
 }
 
@@ -52,7 +58,7 @@ impl<A> Handler<SystemStopMessage> for A
 where
     A: Actor + Sized,
 {
-    fn handle(&mut self, _msg: SystemStopMessage, context: &ActorContext<A>) -> ActorResult {
+    fn handle(&mut self, _msg: SystemStopMessage, context: &ActorContext<A>) -> Result<ActorResult, Box<dyn Error>> {
         return self.on_system_stop(context);
     }
 }
@@ -63,11 +69,11 @@ where
     A: Handler<M>,
     M: ActorMessage,
 {
-    fn handle(&mut self, msg: BulkActorMessage<M>, context: &ActorContext<Self>) -> ActorResult {
+    fn handle(&mut self, msg: BulkActorMessage<M>, context: &ActorContext<Self>) -> Result<ActorResult, Box<dyn Error>> {
         for i in msg.data.into_iter() {
-            self.handle(i, context);
+            self.handle(i, context)?;
         }
-        return ActorResult::Ok;
+        return Ok(ActorResult::Ok);
     }
 }
 
@@ -75,7 +81,7 @@ impl<A> Handler<SerializedMessage> for A
 where
     A: Actor + Sized + Actor,
 {
-    fn handle(&mut self, msg: SerializedMessage, context: &ActorContext<A>) -> ActorResult {
+    fn handle(&mut self, msg: SerializedMessage, context: &ActorContext<A>) -> Result<ActorResult, Box<dyn Error>> {
         return self.handle_serialized_message(msg, context);
     }
 }
@@ -85,7 +91,7 @@ impl<A> Handler<SleepMessage> for A
     where
         A: Actor + Sized + Actor,
 {
-    fn handle(&mut self, msg: SleepMessage, _context: &ActorContext<A>) -> ActorResult {
-        return ActorResult::Sleep(msg.duration)
+    fn handle(&mut self, msg: SleepMessage, _context: &ActorContext<A>) -> Result<ActorResult, Box<dyn Error>> {
+        return Ok(ActorResult::Sleep(msg.duration));
     }
 }