+++ src/lib.rs
@@ -34,9 +34,9 @@
 //! // each supported message has its own Handler implementation
 //! // this is where the actual work is done
 //! impl Handler<FooBar> for HelloWorld {
-//!     fn handle(&mut self, _msg: FooBar, _context: &ActorContext<Self>) -> ActorResult {
+//!     fn handle(&mut self, _msg: FooBar, _context: &ActorContext<Self>) -> Result<ActorResult, Box<dyn Error>> {
 //!         println!("Message Received!");
-//!         ActorResult::Ok
+//!         Ok(ActorResult::Ok)
 //!     }
 //! }
 //!