+++ src/message/envelope.rs
@@ -1,3 +1,4 @@
+use std::error::Error;
 use crate::actor::context::ActorContext;
 use crate::actor::handler::Handler;
 use crate::message::actor_message::ActorMessage;
@@ -7,7 +8,7 @@ pub trait MessageEnvelopeTrait<A>: Send + Sync
 where
     A: Actor,
 {
-    fn handle(&mut self, actor: &mut A, context: &ActorContext<A>) -> ActorResult;
+    fn handle(&mut self, actor: &mut A, context: &ActorContext<A>) -> Result<ActorResult, Box<dyn Error>>;
 }
 
 pub struct MessageEnvelope<A>(Box<dyn MessageEnvelopeTrait<A> + Send + Sync>);
@@ -26,7 +27,7 @@ impl<A> MessageEnvelopeTrait<A> for MessageEnvelope<A>
 where
     A: Actor,
 {
-    fn handle(&mut self, act: &mut A, context: &ActorContext<A>) -> ActorResult {
+    fn handle(&mut self, act: &mut A, context: &ActorContext<A>) -> Result<ActorResult, Box<dyn Error>> {
         return self.0.handle(act, context);
     }
 }
@@ -43,10 +44,10 @@ where
     M: ActorMessage + Send + 'static,
     A: Handler<M> + Actor,
 {
-    fn handle(&mut self, act: &mut A, context: &ActorContext<A>) -> ActorResult {
+    fn handle(&mut self, act: &mut A, context: &ActorContext<A>) -> Result<ActorResult, Box<dyn Error>> {
         if let Some(msg) = self.msg.take() {
             return act.handle(msg, context);
         }
-        return ActorResult::Ok;
+        return Ok(ActorResult::Ok);
     }
 }